modify_incident/requirements.txt
flask==3.0.3
firebase-admin==6.5.0
google-cloud-secret-manager==2.20.2
python-dotenv==1.0.1
gunicorn==23.0.0
google-cloud-firestore==2.18.0